+++ eras/conway/test-suite/test/Test/Cardano/Ledger/Conway/Serialisation/CDDL.hs
@@ -25,6 +25,7 @@ tests :: Int -> TestTree
 tests n = withResource combinedCDDL (const (pure ())) $ \cddl ->
   testGroup "CDDL roundtrip tests" $
     [ cddlTest @(Value Conway) (eraProtVerHigh @Conway) n "coin"
+    , cddlTest @(Value Conway) (eraProtVerHigh @Conway) n "value"
     , cddlAnnotatorTest @(TxBody Conway) (eraProtVerHigh @Conway) n "transaction_body"
     , cddlAnnotatorTest @(TxAuxData Conway) (eraProtVerHigh @Conway) n "auxiliary_data"
     , cddlAnnotatorTest @(Timelock Conway) (eraProtVerHigh @Conway) n "native_script"