+++ libs/cardano-ledger-core/src/Cardano/Ledger/Coin.hs
@@ -19,11 +19,13 @@ module Cardano.Ledger.Coin (
   addDeltaCoin,
   toDeltaCoin,
   integerToWord64,
+  decodePositiveCoin,
 )
 where
 
 import Cardano.HeapWords (HeapWords)
 import Cardano.Ledger.Binary (
+  Decoder,
   FromCBOR (..),
   ToCBOR (..),
   decodeInteger,
@@ -135,3 +137,10 @@ instance ToExpr DeltaCoin
 
 instance ToExpr (CompactForm Coin) where
   toExpr x = toExpr (fromCompact x)
+
+decodePositiveCoin :: Decoder s Coin
+decodePositiveCoin = do
+  n <- decodeWord64
+  if n == 0
+    then fail "Expected a positive Coin. Got 0 (zero)."
+    else pure $ Coin (toInteger n)